Split viewer legacy core helpers
