+++ .pm/tasks/task_f77968a2f95a4b8fa34a542a62a0aec0.yaml
@@ -1,9 +1,9 @@
 task_uid: task_f77968a2f95a4b8fa34a542a62a0aec0
 title: "Finish viewer legacy_core facade split cleanup"
 owner_role: viewer_engineer
-worktree_hint: null
+worktree_hint: /Users/scc/ccwork/worktrees/oasis7-engineering-visual-cleanup-audit
 execution_log_path: .pm/tasks/task_f77968a2f95a4b8fa34a542a62a0aec0.execution.md
-status: candidate
+status: done
 priority: P2
 source_signal: SIG-PM-0065
 source_refs:
@@ -17,4 +17,11 @@ acceptance:
   - "Relevant viewer UI tests and build pass after the split/rebaseline."
 handoff_to:
   - qa_engineer
-updated_at: 2026-06-13T13:14:04+08:00
+updated_at: 2026-06-13T16:19:20+08:00
+last_claim_type: task_complete
+last_verify_command: "npm --prefix crates/oasis7_viewer run test:ui && npm --prefix crates/oasis7_viewer run build:software-safe && node crates/oasis7_viewer/scripts/software-safe-feedback-contract.test.mjs && ./scripts/doc-governance-check.sh && ./scripts/pm/workflow-lint.sh --task-uid task_f77968a2f95a4b8fa34a542a62a0aec0 --phase current && git diff --check"
+last_verified_at: 2026-06-13T16:19:18+08:00
+last_verification_exit_code: 0
+last_verification_status: verified
+last_started_at: 2026-06-13T15:57:40+08:00
+last_closed_at: 2026-06-13T16:00:02+08:00
