+++ crates/oasis7_viewer/software_safe_src/software_safe_state.js
@@ -1,0 +1,101 @@
+import { createMutable } from "solid-js/store";
+import { createInitialHostedLoginState } from "./viewer_hosted_login_state_module.js";
+import { VIEWER_RENDER_MODE } from "./software_safe_constants.js";
+
+export function createSoftwareSafeState() {
+  return createMutable({
+    uiLocale: "en",
+    promptOverridesVisible: false,
+    connectionStatus: "connecting",
+    logicalTime: 0,
+    eventSeq: 0,
+    tick: 0,
+    selectedKind: null,
+    selectedId: null,
+    errorCount: 0,
+    lastError: null,
+    eventCount: 0,
+    traceCount: 0,
+    cameraMode: "viewer",
+    cameraRadius: 0,
+    cameraOrthoScale: 0,
+    renderMode: VIEWER_RENDER_MODE,
+    rendererClass: "none",
+    viewerReason: null,
+    renderer: null,
+    vendor: null,
+    webglVersion: null,
+    pixelWorldRuntimeStatus: "detached",
+    pixelWorldRuntimeSource: "detached",
+    pixelWorldRuntimeModuleUrl: null,
+    pixelWorldCamera: null,
+    pixelWorldFatal: null,
+    controlProfile: "playback",
+    worldId: null,
+    server: null,
+    wsUrl: null,
+    lastControlFeedback: null,
+    lastPromptFeedback: null,
+    lastChatFeedback: null,
+    lastGameplayActionFeedback: null,
+    snapshot: null,
+    metrics: null,
+    hostedAccess: null,
+    hostedAdmission: null,
+    recentEvents: [],
+    recentDecisionTraces: [],
+    chatHistory: [],
+    selectedObject: null,
+    auth: {
+      available: false,
+      hostedAccountId: null,
+      playerId: null,
+      loginChannel: null,
+      maskedLoginHint: null,
+      deviceSessionId: null,
+      publicKey: null,
+      privateKey: null,
+      releaseToken: null,
+      error: null,
+      revokeReason: null,
+      revokedBy: null,
+      source: "guest_only",
+      registrationStatus: "guest",
+      sessionEpoch: null,
+      issuedAtUnixMs: null,
+      recoveryErrorCode: null,
+      recoveryErrorMessage: null,
+      issueInFlight: false,
+      syncInFlight: false,
+      runtimeStatus: "guest",
+      boundAgentId: null,
+      pendingRequestedAgentId: null,
+      pendingForceRebind: false,
+      rebindNotice: null,
+    },
+    hostedLogin: createInitialHostedLoginState(),
+    promptDraft: {
+      agentId: null,
+      currentVersion: 0,
+      rollbackTargetVersion: 0,
+      updatedBy: "",
+      updatedAtTick: 0,
+      systemPrompt: "",
+      shortTermGoal: "",
+      longTermGoal: "",
+      dirty: false,
+    },
+    chatDraft: {
+      agentId: null,
+      message: "",
+      dirty: false,
+    },
+    strongAuth: {
+      approvalCode: "",
+      lastGrantActionId: null,
+      lastGrantExpiresAtUnixMs: null,
+      lastGrantError: null,
+    },
+    selectedSearch: "",
+  });
+}
