+++ crates/oasis7_viewer/software_safe_src/viewer_auth_crypto.js
@@ -1,0 +1,178 @@
+import { VIEWER_AUTH_SIGNATURE_PREFIX } from "./software_safe_constants.js";
+
+const ED25519_PKCS8_PREFIX = new Uint8Array([
+  0x30, 0x2e, 0x02, 0x01, 0x00, 0x30, 0x05, 0x06,
+  0x03, 0x2b, 0x65, 0x70, 0x04, 0x22, 0x04, 0x20,
+]);
+const textEncoder = new TextEncoder();
+const authKeyCache = new Map();
+
+function cborHeader(majorType, length) {
+  if (!Number.isInteger(length) || length < 0) {
+    throw new Error(`invalid CBOR length: ${length}`);
+  }
+  if (length < 24) {
+    return Uint8Array.of((majorType << 5) | length);
+  }
+  if (length < 0x100) {
+    return Uint8Array.of((majorType << 5) | 24, length);
+  }
+  if (length < 0x10000) {
+    return Uint8Array.of((majorType << 5) | 25, (length >> 8) & 0xff, length & 0xff);
+  }
+  if (length <= 0xffffffff) {
+    return Uint8Array.of(
+      (majorType << 5) | 26,
+      (length >>> 24) & 0xff,
+      (length >>> 16) & 0xff,
+      (length >>> 8) & 0xff,
+      length & 0xff,
+    );
+  }
+  if (length <= Number.MAX_SAFE_INTEGER) {
+    const value = BigInt(length);
+    return Uint8Array.of(
+      (majorType << 5) | 27,
+      Number((value >> 56n) & 0xffn),
+      Number((value >> 48n) & 0xffn),
+      Number((value >> 40n) & 0xffn),
+      Number((value >> 32n) & 0xffn),
+      Number((value >> 24n) & 0xffn),
+      Number((value >> 16n) & 0xffn),
+      Number((value >> 8n) & 0xffn),
+      Number(value & 0xffn),
+    );
+  }
+  throw new Error("CBOR length exceeds Number.MAX_SAFE_INTEGER");
+}
+
+function concatBytes(...parts) {
+  const totalLength = parts.reduce((sum, bytes) => sum + bytes.length, 0);
+  const out = new Uint8Array(totalLength);
+  let offset = 0;
+  for (const bytes of parts) {
+    out.set(bytes, offset);
+    offset += bytes.length;
+  }
+  return out;
+}
+
+export function cborEncode(value) {
+  if (value === null) {
+    return Uint8Array.of(0xf6);
+  }
+  if (value === false) {
+    return Uint8Array.of(0xf4);
+  }
+  if (value === true) {
+    return Uint8Array.of(0xf5);
+  }
+  if (typeof value === "number") {
+    if (!Number.isInteger(value) || value < 0) {
+      throw new Error(`unsupported CBOR number: ${value}`);
+    }
+    return cborHeader(0, value);
+  }
+  if (typeof value === "string") {
+    const bytes = textEncoder.encode(value);
+    return concatBytes(cborHeader(3, bytes.length), bytes);
+  }
+  if (Array.isArray(value)) {
+    return concatBytes(cborHeader(4, value.length), ...value.map((entry) => cborEncode(entry)));
+  }
+  if (value instanceof Uint8Array) {
+    return concatBytes(cborHeader(2, value.length), value);
+  }
+  if (typeof value === "object") {
+    const entries = Object.entries(value).filter(([, entryValue]) => entryValue !== undefined);
+    const encoded = [cborHeader(5, entries.length)];
+    for (const [key, entryValue] of entries) {
+      encoded.push(cborEncode(String(key)));
+      encoded.push(cborEncode(entryValue));
+    }
+    return concatBytes(...encoded);
+  }
+  throw new Error(`unsupported CBOR type: ${typeof value}`);
+}
+
+function hexToBytes(raw) {
+  const value = String(raw || "").trim().toLowerCase();
+  if (!value || value.length % 2 !== 0 || /[^0-9a-f]/.test(value)) {
+    throw new Error("invalid hex payload");
+  }
+  const bytes = new Uint8Array(value.length / 2);
+  for (let index = 0; index < bytes.length; index += 1) {
+    bytes[index] = Number.parseInt(value.slice(index * 2, index * 2 + 2), 16);
+  }
+  return bytes;
+}
+
+function bytesToHex(bytes) {
+  return Array.from(bytes, (value) => value.toString(16).padStart(2, "0")).join("");
+}
+
+function bytesStartWith(bytes, prefix) {
+  if (bytes.length < prefix.length) {
+    return false;
+  }
+  for (let index = 0; index < prefix.length; index += 1) {
+    if (bytes[index] !== prefix[index]) {
+      return false;
+    }
+  }
+  return true;
+}
+
+async function importEd25519SigningKey(privateKeyHex) {
+  if (!window.crypto?.subtle) {
+    throw new Error("Web Crypto subtle API is unavailable");
+  }
+  if (!authKeyCache.has(privateKeyHex)) {
+    const rawPrivateKey = hexToBytes(privateKeyHex);
+    if (rawPrivateKey.length !== 32) {
+      throw new Error(`viewer auth private key length mismatch: expected 32 bytes, got ${rawPrivateKey.length}`);
+    }
+    const pkcs8 = concatBytes(ED25519_PKCS8_PREFIX, rawPrivateKey);
+    authKeyCache.set(
+      privateKeyHex,
+      window.crypto.subtle.importKey("pkcs8", pkcs8, { name: "Ed25519" }, false, ["sign"]),
+    );
+  }
+  return authKeyCache.get(privateKeyHex);
+}
+
+export async function signAuthPayload(signingPayloadBytes, auth) {
+  const key = await importEd25519SigningKey(auth.privateKey);
+  const signature = await window.crypto.subtle.sign({ name: "Ed25519" }, key, signingPayloadBytes);
+  return `${VIEWER_AUTH_SIGNATURE_PREFIX}${bytesToHex(new Uint8Array(signature))}`;
+}
+
+export async function generateEphemeralEd25519Keypair() {
+  if (!window.crypto?.subtle) {
+    throw new Error("Web Crypto subtle API is unavailable");
+  }
+  const keyPair = await window.crypto.subtle.generateKey(
+    { name: "Ed25519" },
+    true,
+    ["sign", "verify"],
+  );
+  const pkcs8 = new Uint8Array(await window.crypto.subtle.exportKey("pkcs8", keyPair.privateKey));
+  if (!bytesStartWith(pkcs8, ED25519_PKCS8_PREFIX) || pkcs8.length !== ED25519_PKCS8_PREFIX.length + 32) {
+    throw new Error("unexpected Ed25519 pkcs8 encoding from Web Crypto");
+  }
+  const rawPublicKey = new Uint8Array(await window.crypto.subtle.exportKey("raw", keyPair.publicKey));
+  if (rawPublicKey.length !== 32) {
+    throw new Error(`unexpected Ed25519 public key length: ${rawPublicKey.length}`);
+  }
+  return {
+    publicKey: bytesToHex(rawPublicKey),
+    privateKey: bytesToHex(pkcs8.slice(ED25519_PKCS8_PREFIX.length)),
+  };
+}
+
+export function buildAuthEnvelope(payload) {
+  return cborEncode({
+    version: 1,
+    payload,
+  });
+}
